fix bug

+++ app/Http/Controllers/Back/CategoryController.php
@@ -71,12 +71,10 @@ class CategoryController extends Controller
     {
         $post = self::fileUpload($request);
         if(!$post['result']){
-            return redirect()->to($this->getRedirectUrl())
-                    ->withInput($request->input())
-                    ->withErrors($post['error'], $this->errorBag() );
+            return redirectBack($post['error'],$this->errorBag(),$this->getRedirectUrl(),$request->input());
         }
         category::create($post);
-        return redirect($this->route);
+        return redirectFlash('เพิ่มรายการ สำเร็จ',$this->route);
     }
 
     /**
@@ -119,16 +117,10 @@ class CategoryController extends Controller
     {
         $post = self::fileUpload($request);
         if(!$post['result']){
-            return redirect()->to($this->getRedirectUrl())
-                    ->withInput($request->input())
-                    ->withErrors($post['error'], $this->errorBag() );
+            return redirectBack($post['error'],$this->errorBag(),$this->getRedirectUrl(),$request->input());
         }
-        // if ($post['position']==1){
-        //     // $post['position'] = 999;
-        // }
         $db = category::find($id)->update($post) ;
-        session()->flash('message','Updated Successfully');
-        return redirect($this->route);
+        return redirectFlash('อัพเดทรายการ สำเร็จ',$this->route);
     }
 
     /**
@@ -139,9 +131,11 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        if(Auth::user()->id != 1 ){
+            return redirectFlash('ไม่สามารถลบรายการนี้ได้ <BR> เนื่องจากไอดีของคุณ ไม่มี permission ค่ะ',$this->route,'error');
+        }
         category::find($id)->delete();
-        session()->flash('message','Delete Successfully');
-        return redirect($this->route);
+        return redirectFlash('ลบรายการ สำเร็จ',$this->route);
     }
